Simplify specifying a job by letting submit read a job description from stdin and spawns the lambdas; now mk-wc-job doesn't have to deal with json.
Handle pair-wise dependencies better so that the producer doesn't
disappear until the consumer is running and until the consumer has
finished.

Keep tracker of nreader/nwriter for a pipe so that read/write fails if there
is no corresponding writer/reader.

Rename the ulambdad package to schedd.  The package itself needs some cleanup.