Improve paging through directories
Introduce a cursor to keep track how much of a directory has been
read.  The cursor is in terms of dirents and allows a better
implementation for paging through a directory.

Add test for paging through a directory

Make CHUNKSZ a field of fslib so that we can set dynamically (e.g., in
tests)