Fix race with Watch and inode version numbers
Lookup() returned a slice with inode pointers and then Walk() computed
the Qids to return for for each inode. But, between Lookup() returning
and computing Qids, a Remove() can update the version # of the inode.
Then, a later Watch() won't fail because the client will provide the
version # from after the remove instead of from before the remove.
Fix Lookup to return Qids and compute the Qid while holding the
directory and inode lock.