Retry Spawn if groupmgr fails to Spawn
The manager may a procd mounted that crashed and respawn member i at
the procd before the ephemeral file of that procd has been
removed. Without retrying, the manager's walk of procd/~ip/ will
umount the failed procd but walk might be fast and find the dead
procd's symlink and mount it, and then fail to write to its ctl file.