Two bugs fixes:
- walkPath may return a longer left path than started with because of
mount points that point to a long path; we shouldn't use the original
path for umount on failure, but the newer longer found path.

- disconnect should umount path to guarantee that the server will not
be remounted, which reflects the intention of testing with partitions.

make the caller of umount free the fid so that disconnect can call
disconnect on it first.