Checkpoint WIP on using protobuf for Rattach
This change is invasive because Rattach returns a qid, which has a
type field of uint8, but protobuf supports uint32 as the smallest
unit.  Thus, we have two qids: one for sigmap and one for 9P, which
means also two Stat's, which means that the proxy must convert between
sigmaP and nineP structs when reading a directory.