Delete version numbers from Read and Write
The only use case was updating a file with an epoch number but now
epoch numbers come etcd and leader election allow a client to "own" a
file.

Delete the TestCounter, which tests ReadV/WriteV.

Small benefit: pipe tests from memfs work again. It is not ok to hold
a file lock across a blocking read.