Fix race between updating the directory cache inside etcd
fsetcd assumes caller holds lock on the directory but ReadF doesn't
take a lock, but ReadF maybe reading a directory and thus may update
the directory cache. If it reads the directory before a create in that
directory and inserts it in the cache after create succeeded, then it
inserts a stale entry, causing TestConcurDir to fail.