Checkpoint WIP on binfsd
uprocsrv starts binfsd, which is a fuse server that serves binaries to
the Linux kernel; it mounts itself in /mnt/binfs.  uprocsrv execs a
binary with a pathname inside /mnt/binfs.  if binfsd doesn't have the
binary in its file system, it will download it to its local file
system and then serve it to the kernel. the potential benefit of this
design is that sigmaos can leverage the kernel's demand paging.

Passes TestWaitExitSimpleSingleBE