Implement a simpler and more correct plan for handling expired ephemeral files in fsetcd.
This plan stores the pathname with the ephemeral key of an ephemeral
file in etcd on Create. Rename updates the ephemeral key and Remove
deletes it, respectively.  This plan provides a reverse index and
doesn't rely on searching dcache entries (which may not exist yet, if
the watch expires before reading the directory into dcache).