If WaitExit() fails, shutdown named
If WaitExit() losts its connection to its msched, it might as well
fail, and rely on groupmgr to start a new named for this realm.  This
shortens recovery; otherwise, the new named cannot start until all
WaitExit() attempts have failed plus lease timeout.

The loop might be a left-over from when the kernel transitioned
rootrealm knamed to a named and then back.